requirements.txt
asgiref==3.3.1
autopep8==1.5.5
Django==3.1.7
django-ckeditor==6.0.0
django-crispy-forms==1.11.1
django-js-asset==1.2.2
freeze==3.0
mysqlclient==2.0.3
Pillow==8.1.1
pycodestyle==2.6.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
